Implement shaders variants
Very inneficient for now, but testing shows great accuracy improvements. Basically, instead of assuming the pairs of shaders (which might not be as simple as one vert, one frag), we generate multi_compile and shader_feature directives with all the keywords combinations and put a giant if elif preprocessor chain to selectively enable the shader with the correct combinations.

This mapping isn't 1 to 1: ShaderLab only allows to declare sets of keywords, but we want specific combinations. As a result, this algorithm can overgenerate variants, but it is guaranteed to always include the ones we need.

It also completely changes the structure: we now have a single basket with all shaders and we go through them sorted by type, then by most specific in keywords to least specific. This lets us have a fallback else directive on the last one while making sure all the more specific combinations were checked.

More optimisations will be needed to merge the variants together: this can make the files significantly larger (thousands of lines more in worst cases), but for now, this works